requirements.txt
ipython
psycopg2-binary
typer
python-dateutil
